Move test providers to dedicated folder
